+++ pkg/connector/commands.go
@@ -17,12 +17,15 @@
 package connector
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"html"
+	"slices"
 	"strings"
 
 	"github.com/rs/zerolog"
+	"go.mau.fi/util/exslices"
 	"go.mau.fi/whatsmeow"
 	"go.mau.fi/whatsmeow/appstate"
 	"go.mau.fi/whatsmeow/types"
@@ -120,9 +123,6 @@ func fnSync(ce *commands.Event) {
 			return
 		}
 		for _, group := range groups {
-			wrapped := wa.wrapGroupInfo(ce.Ctx, group)
-			wrapped.ExtraUpdates = bridgev2.MergeExtraUpdaters(wrapped.ExtraUpdates, updatePortalLastSyncAt)
-			wa.addExtrasToWrapped(ce.Ctx, group.JID, wrapped, nil)
 			login.QueueRemoteEvent(&simplevent.ChatResync{
 				EventMeta: simplevent.EventMeta{
 					Type:         bridgev2.RemoteEventChatResync,
@@ -130,7 +130,12 @@ func fnSync(ce *commands.Event) {
 					LogContext:   logContext,
 					CreatePortal: true,
 				},
-				ChatInfo: wrapped,
+				GetChatInfoFunc: func(ctx context.Context, portal *bridgev2.Portal) (*bridgev2.ChatInfo, error) {
+					wrapped := wa.wrapGroupInfo(ce.Ctx, group)
+					wrapped.ExtraUpdates = bridgev2.MergeExtraUpdaters(wrapped.ExtraUpdates, updatePortalLastSyncAt)
+					wa.addExtrasToWrapped(ce.Ctx, group.JID, wrapped, nil, portal.MXID == "")
+					return wrapped, nil
+				},
 			})
 		}
 		ce.Reply("Queued syncs for %d groups", len(groups))
@@ -141,7 +146,17 @@ func fnSync(ce *commands.Event) {
 		wa.resyncContacts(true, false)
 		ce.React("✅")
 	case "appstate":
-		for _, name := range appstate.AllPatchNames {
+		names := appstate.AllPatchNames[:]
+		if len(ce.Args) > 1 {
+			names = exslices.CastFuncFilter(ce.Args[1:], func(name string) (appstate.WAPatchName, bool) {
+				if !slices.Contains(appstate.AllPatchNames[:], appstate.WAPatchName(name)) {
+					ce.Reply("Invalid app state name `%s`", name)
+					return "", false
+				}
+				return appstate.WAPatchName(name), true
+			})
+		}
+		for _, name := range names {
 			err := wa.Client.FetchAppState(ce.Ctx, name, true, false)
 			if errors.Is(err, appstate.ErrKeyNotFound) {
 				ce.Reply("Key not found error syncing app state %s: %v\n\nKey requests are sent automatically, and the sync should happen in the background after your phone responds.", name, err)