+++ pkg/connector/events.go
@@ -133,6 +133,14 @@ func (evt *WAMessageEvent) PreHandle(ctx context.Context, portal *bridgev2.Porta
 		return
 	}
 	meta := portal.Metadata.(*waid.PortalMetadata)
+	if meta.AddressingMode == types.AddressingModeLID && evt.Info.Sender.Server == types.DefaultUserServer {
+		evt.Info.Sender, evt.Info.SenderAlt = evt.Info.SenderAlt, evt.Info.Sender
+		zerolog.Ctx(ctx).Debug().
+			Stringer("lid", evt.Info.Sender).
+			Stringer("pn", evt.Info.SenderAlt).
+			Str("message_id", evt.Info.ID).
+			Msg("Forced phone number sender to LID in group message")
+	}
 	if meta.AddressingMode == types.AddressingModeLID || meta.LIDMigrationAttempted {
 		return
 	}
@@ -209,9 +217,15 @@ func (evt *WAMessageEvent) ConvertEdit(ctx context.Context, portal *bridgev2.Por
 
 func (evt *WAMessageEvent) GetTargetMessage() networkid.MessageID {
 	if reactionMsg := evt.Message.GetReactionMessage(); reactionMsg != nil {
-		return msgconv.KeyToMessageID(evt.wa.Client, evt.Info.Chat, evt.Info.Sender, reactionMsg.GetKey())
+		ctx := evt.wa.UserLogin.Log.
+			With().Str("action", "get reaction target message").Str("message_id", evt.Info.ID).Logger().
+			WithContext(evt.wa.Main.Bridge.BackgroundCtx)
+		return msgconv.KeyToMessageID(ctx, evt.wa.Client, evt.Info.Chat, evt.Info.Sender, reactionMsg.GetKey())
 	} else if protocolMsg := evt.Message.GetProtocolMessage(); protocolMsg != nil {
-		return msgconv.KeyToMessageID(evt.wa.Client, evt.Info.Chat, evt.Info.Sender, protocolMsg.GetKey())
+		ctx := evt.wa.UserLogin.Log.
+			With().Str("action", "get edit target message").Str("message_id", evt.Info.ID).Logger().
+			WithContext(evt.wa.Main.Bridge.BackgroundCtx)
+		return msgconv.KeyToMessageID(ctx, evt.wa.Client, evt.Info.Chat, evt.Info.Sender, protocolMsg.GetKey())
 	}
 	return ""
 }