+++ pkg/waid/dbmeta.go
@@ -20,10 +20,12 @@ import (
 	"crypto/ecdh"
 	"crypto/rand"
 	"encoding/json"
+	"time"
 
 	"go.mau.fi/util/exerrors"
 	"go.mau.fi/util/jsontime"
 	"go.mau.fi/util/random"
+	"go.mau.fi/whatsmeow/appstate"
 	"go.mau.fi/whatsmeow/types"
 )
 
@@ -37,7 +39,11 @@ type UserLoginMetadata struct {
 	APNSEncPrivKey  []byte        `json:"apns_enc_privkey,omitempty"`
 	LoggedInAt      jsontime.Unix `json:"logged_in_at,omitempty"`
 
+	AppStateRecoveryAttempted map[appstate.WAPatchName]time.Time `json:"app_state_recovery_attempted,omitempty"`
+
 	HistorySyncPortalsNeedCreating bool `json:"history_sync_portals_need_creating,omitempty"`
+
+	MData json.RawMessage `json:"mdata,omitempty"`
 }
 
 type PushKeys struct {