+++ CHANGELOG.md
@@ -1,3 +1,48 @@
+# v26.02
+
+* Bumped minimum Go version to 1.25.
+* Added automatic recovery for WhatsApp app state sync issues.
+* Fixed LID redirects for some non-message events.
+
+# v26.01
+
+* Fixed broadcast list messages to LIDs causing split DMs.
+* Fixed read receipts not working correctly in LID DMs.
+* Fixed backfill sometimes racing with receiving LID mappings.
+
+# v25.12
+
+* Updated Docker image to Alpine 3.23.
+* Fixed group member invites from Matrix not automatically disinviting the phone
+  number ghost when the invite is redirected to a LID ghost.
+
+# v25.11
+
+* Added interface support for notifying about failed invites when creating a
+  group and sending the invites via DM (only applicable to provisioning API).
+* Added migration to automatically delete duplicate LID DM portals that were
+  created earlier.
+* Changed contact list API to only include actual phone contacts.
+* Removed extra unrecognized message notice when receiving live photos
+  (bridging the live photo video is not currently planned).
+* Fixed pairing not working with latest WhatsApp Android version.
+* Fixed replies, read receipts and typing notifications not being bridged
+  correctly after DM LID migration.
+* Fixed backfill creating duplicate portals if history sync contains both LID
+  and phone number DM data.
+* Fixed some cases of LID and phone number user infos getting out of sync.
+* Fixed muting chat forever not being bridged correctly from WhatsApp.
+* Fixed old mutes being re-applied on chat resync in some cases.
+* Fixed backfilling failing if some reactions were missing sender info.
+* Fixed space not being deleted when leaving community on WhatsApp.
+* Fixed sticker size metadata on Matrix not matching how native WhatsApp Web
+  renders them.
+* Fixed ratelimit errors in login not being exposed to the user properly
+  (thanks to [@dead8309] in [#852]).
+
+[@dead8309]: https://github.com/dead8309
+[#852]: https://github.com/mautrix/whatsapp/pull/852
+
 # v25.10
 
 * Switched to calendar versioning.